[Rename] comp 내 파일 molcules/organisms로 분리
