+++ src/App.js
@@ -3,9 +3,8 @@ import Login from './pages/Login/Login';
 import EmailLogin from './pages/EmailLogin/EmailLogin';
 import Splash from './pages/Splash/Splash';
 import Home from './pages/Home/Home';
-import YourProfile from './pages/Profile/YourProfile';
+import YourProfile from './pages/Profile/YourProfile/YourProfile';
 import Chat from './pages/Chat/Chat';
-import Posts from './components/molecules/Posts/Posts';
 
 function App() {
     return (
