+++ src/components/molecules/Empty/Empty.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import defaultTheme from '../../commons/style/themes/default';
+import defaultTheme from '../../../commons/style/themes/default';
 
 export const Empty = styled.div`
     background-color: ${defaultTheme.palette.gray};
